Fix the last compile errors: the build is green
Two errors were left, both pre-existing and both invisible until Lombok
was restored to the annotation processor path.

extraPrice was a double
-----------------------
ProductOptionValue held a surcharge as a primitive double while
ProductRequest declared it a BigDecimal, and the request's getter
narrowed the field back to double — which is what did not compile, and
what would have silently rounded a surcharge on its way in.

The entity now holds BigDecimal(12,2), like every other money column in
the model, and the getter returns the type of the field it reads. The
request field is optional, so a value with no surcharge is stored as zero
rather than null, and getEffectivePrice skips a null surcharge rather
than failing the whole product.

Vendor.acive
------------
A @Builder.Default with no initializer, which made Lombok emit a call to
a $default$acive() it never generated. Removed rather than defaulted: no
code ever set it, so `if (!vendor.getAcive())` in requireActiveVendor
unboxed a null and would have thrown on every product create. Vendor.
status already answers whether a vendor may trade, so that guard now
tests APPROVED or ACTIVE — the same rule order placement applies — and
names the offending status when it refuses.

Also added a placeholder in src/main/resources/graphql-client: the DGS
codegen plugin scans that path and fails the build when it is absent, so
a fresh clone could not compile even with the source correct. Delete it
along with the plugin if the GraphQL client is not coming back.

mvn compile and mvn test both pass: 39 tests, no failures.

+++ src/main/java/com/sujula/dto/request/product/ProductRequest.java
@@ -96,7 +96,7 @@ public class ProductRequest {
 
         public String getValue() { return value; }
         public void setValue(String value) { this.value = value; }
-        public double getExtraPrice() { return extraPrice; }
+        public BigDecimal getExtraPrice() { return extraPrice; }
         public void setExtraPrice(BigDecimal extraPrice) { this.extraPrice = extraPrice; }
     }
